[se edito el endpoints delete invetory]

+++ src/app.py
@@ -16,6 +16,7 @@ from flask_jwt_extended import create_access_token
 from flask_jwt_extended import get_jwt_identity
 from flask_jwt_extended import jwt_required
 from flask_jwt_extended import JWTManager
+from flask_cors import CORS
 from flask_bcrypt import Bcrypt
 
 # from models import Person
@@ -28,6 +29,7 @@ app.url_map.strict_slashes = False
 app.config["JWT_SECRET_KEY"] = os.getenv("JWT_KEY")
 jwt = JWTManager(app)
 bcrypt = Bcrypt(app)
+CORS(app)
 
 # database condiguration
 db_url = os.getenv("DATABASE_URL")
@@ -373,10 +375,16 @@ def update_inventory_item(id):
 
 #-- Verificar token y mirar si el id del inventario le corresponde a la compañia del token-- ##
 @app.route('/inventory/<int:id>', methods=['DELETE'])
+@jwt_required()
 def delete_inventory_item(id):
+    company_id = get_jwt_identity()
+
     item = db.session.get(Inventory, id)
     if not item:
         return jsonify({'msg': 'Item not found'}), 404
+
+    if item.companyID != company_id:
+        return jsonify({'msg': 'Unauthorized to delete this item'}), 404
     db.session.delete(item)
     db.session.commit()
     return jsonify({'msg': 'Item deleted'}), 200
